+++ utility/redis.go
@@ -6,13 +6,13 @@ import (
 	"strconv"
 )
 
-func RedisHSet(serviceName string, param string, value int) {
+func RedisHSet(serviceName string, param string, value int64) {
 	client := settings.GetRedisClient()
 
 	client.HSet(context.Background(), serviceName, param, value)
 }
 
-func RedisHGet(serviceName string, param string) int {
+func RedisHGet(serviceName string, param string) int64 {
 	client := settings.GetRedisClient()
 	val, err := client.HGet(context.Background(), serviceName, param).Result()
 	if err != nil {
@@ -23,5 +23,5 @@ func RedisHGet(serviceName string, param string) int {
 	if err != nil {
 		return 0
 	}
-	return i
+	return int64(i)
 }
